requirements.txt
attrs==23.2.0 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.9" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0" and implementation_name == "pypy"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
dotty-dict==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
flywheel-bids @ git+https://gitlab.com/flywheel-io/public/bids-client.git@6ac6a8df4624ea0593c2b2fdc3e1013ec32ac9d0 ; python_version >= "3.9" and python_version < "4.0"
flywheel-gear-toolkit==0.6.18 ; python_version >= "3.9" and python_version < "4.0"
flywheel-gears==0.3.1 ; python_version >= "3.9" and python_version < "4.0"
flywheel-sdk==16.19.2 ; python_version >= "3.9" and python_full_version < "3.10.0"
flywheel-sdk==17.8.0 ; python_full_version >= "3.10.0" and python_version < "4.0"
future==0.18.3 ; python_version >= "3.9" and python_version < "4.0"
idna==3.7 ; python_version >= "3.9" and python_version < "4.0"
importlib-resources==6.4.0 ; python_version >= "3.9" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.9" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.9" and python_version < "4.0"
psutil==5.9.8 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.9" and python_version < "4.0" and implementation_name == "pypy"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
pyzmq==25.1.2 ; python_version >= "3.9" and python_version < "4.0"
referencing==0.34.0 ; python_version >= "3.9" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rfc3987==1.3.8 ; python_version >= "3.9" and python_version < "4.0"
rpds-py==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
rtstatlib==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.18.1 ; python_version >= "3.9" and python_version < "3.10"
